split files, create contexis
